Switch to a bison/flex combo for parsing/lexing.
Removes the ad-hoc parser/lexer that used getchar heavily.
As a result, the main.cpp file is simplified greatly and opens
up the opportunity to factor out the existing line-reading stuff
to a REPL setup and have main use that or read from a file depending
on how it's invoked.

KNOWN ISSUES:

* gotta either re-introduce the custom operator syntax or
  get the example ones into the grammar
* recursion (sometimes) leads to a segfault when evaluating the
  func_pointer in main.cpp
* some spurious syntax errors when ending execution, probably
  a straggling EOF token or something
* whenever multi-line expression is entered, there are a handful
  of blank lines printed out, # of blank lines is same # of
  expression lines.